Update TCP connection tracking example
Change it so it uses a ternary P4 table instead of an if-then-else-if
statement.  This is sometimes used in P4 targets with TCAMs since it
can use less hardware resources than an if-then-else-if statement.

Remove the call to allocate_flow_id() since its return value was not
needed, nor used, for this example.