Replace 'inline extern' with 'fixed function block' The term 'inline extern' is a misleading choice of words for this functionality in PNA, and fixed function block has a longer history of being used for such things.
Remove a left-over mention of the pre control that was missed before.

Correct the name of the extern function set_entry_expire_time_if in
some examples in comments.

Use INCLUDE directive in a few more places of PNA.mdk that will help
keep the code snippets from the pna.p4 include file consistent between
that file and the spec more easily in the future.